feat: add auth on frontend and backend
