+++ apps/CEventService.API/Program.cs
@@ -123,6 +123,7 @@ if (app.Environment.IsDevelopment())
 
 app.UseCors("AllowAllOrigins");
 
+app.UseAuthentication();
 app.UseAuthorization();
 app.MapControllers();
