+++ apps/cevent_frontend/src/app/profile/page.tsx
@@ -1,7 +1,10 @@
+import AuthStatus from "@/components/AuthStatus";
+
 export default function ProfilePage() {
   return (
     <div>
       <h1>Profile</h1>
+      <AuthStatus />
     </div>
   );
 }
